\name{ropenphacts-package}
\alias{ropenphacts-package}
\alias{ropenphacts}
\docType{package}
\title{
Functionality to work with the Open PHACTS platform.
}
\description{
Methods to query data from the Open PHACTS platform.
}
\details{
Package based on the ops4j library to make R a client
for the Open PHACTS platform.
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ RDF }
\keyword{ Open PHACTS }
